Move new utility methods to non-public API
