+++ src/main/java/org/apache/maven/buildcache/xml/DtoUtils.java
@@ -20,8 +20,6 @@ package org.apache.maven.buildcache.xml;
 
 import javax.annotation.Nonnull;
 
-import java.io.File;
-import java.nio.file.FileSystems;
 import java.nio.file.Path;
 import java.util.List;
 
@@ -33,11 +31,7 @@ import org.apache.maven.buildcache.xml.build.CompletedExecution;
 import org.apache.maven.buildcache.xml.build.DigestItem;
 import org.apache.maven.buildcache.xml.build.PropertyValue;
 import org.apache.maven.buildcache.xml.config.TrackedProperty;
-import org.apache.maven.execution.MavenSession;
 import org.apache.maven.model.Dependency;
-import org.apache.maven.plugin.MojoExecution;
-import org.apache.maven.plugin.PluginParameterExpressionEvaluator;
-import org.codehaus.plexus.component.configurator.expression.ExpressionEvaluationException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -116,20 +110,20 @@ public class DtoUtils {
     }
 
     /**
-     * @deprecated use {@link #addProperty(CompletedExecution, String, Object, Path, boolean)}
+     *
+     * @deprecated Kept for binary compatibility, not used by the codebase anymore.
+     *             Use {@link CacheUtils#addProperty(CompletedExecution, String, Object, Path, boolean)}
      */
     @Deprecated
     public static void addProperty(
             CompletedExecution execution, String propertyName, Object value, String baseDirPath, boolean tracked) {
-        addProperty(execution, propertyName, value, FileSystems.getDefault().getPath(baseDirPath), tracked);
-    }
-
-    public static void addProperty(
-            CompletedExecution execution, String propertyName, Object value, Path baseDirPath, boolean tracked) {
         final PropertyValue valueType = new PropertyValue();
         valueType.setName(propertyName);
-        final String valueText = normalizeValue(value, baseDirPath);
-        valueType.setValue(valueText);
+        if (value != null && value.getClass().isArray()) {
+            value = ArrayUtils.toString(value);
+        }
+        final String valueText = String.valueOf(value);
+        valueType.setValue(Strings.CS.remove(valueText, baseDirPath));
         valueType.setTracked(tracked);
         execution.addProperty(valueType);
     }
@@ -185,56 +179,5 @@ public class DtoUtils {
         copy.setFileHash(artifact.getFileHash());
         copy.setFileSize(artifact.getFileSize());
         return copy;
-    }
-
-    public static String normalizeValue(Object value, Path baseDirPath) {
-        if (value instanceof File) {
-            Path path = ((File) value).toPath();
-            return normalizedPath(path, baseDirPath);
-        } else if (value instanceof Path) {
-            return normalizedPath(((Path) value), baseDirPath);
-        } else if (value != null && value.getClass().isArray()) {
-            return ArrayUtils.toString(value);
-        } else {
-            return String.valueOf(value);
-        }
-    }
-
-    /*
-     Best effort to normalize paths from Mojo fields.
-     - all absolute paths under project root are relativized for portability
-     - redundant '..' and '.' are removed to have consistent views on all paths
-     - all relative paths are considered portable and are not be touched
-     - absolute paths outside of project directory cannot be deterministically
-       relativized and are not touched
-    */
-    private static String normalizedPath(Path path, Path baseDirPath) {
-        boolean isProjectSubdir = path.isAbsolute() && path.startsWith(baseDirPath);
-        if (LOGGER.isDebugEnabled()) {
-            LOGGER.debug(
-                    "normalizedPath isProjectSubdir {} path '{}' - baseDirPath '{}', path.isAbsolute() {},"
-                            + " path.startsWith(baseDirPath) {}",
-                    isProjectSubdir,
-                    path,
-                    baseDirPath,
-                    path.isAbsolute(),
-                    path.startsWith(baseDirPath));
-        }
-        Path preparedPath = isProjectSubdir ? baseDirPath.relativize(path) : path;
-        String normalizedPath = preparedPath.normalize().toString();
-        if (LOGGER.isDebugEnabled()) {
-            LOGGER.debug("normalizedPath '{}' - {} return {}", path, baseDirPath, normalizedPath);
-        }
-        return normalizedPath;
-    }
-
-    public static Object interpolateExpression(String expression, MavenSession session, MojoExecution execution) {
-        try {
-            PluginParameterExpressionEvaluator evaluator = new PluginParameterExpressionEvaluator(session, execution);
-            return evaluator.evaluate(expression);
-        } catch (ExpressionEvaluationException e) {
-            LOGGER.warn("Cannot interpolate expression '{}': {}", expression, e.getMessage(), e);
-            return expression; // return the expression as is when interpolation fails
-        }
     }
 }
